Copying     + 添加libogg以及libvorbis的版权信息
Engine
    * 添加部分用于linux的代码

    + AudioSourceObject
      音源对象

    * Camera
      + 添加offset matrix用于摄像机震动效果

    + CameraShakeEffect
      摄像机震动效果

    * Engine
      * 初始化流程稍作修改，将其他系统的初始化移到Engine::Initialize()中完成

    * IRenderer
      + 添加RenderTarget相关接口

    + QuadTree
      + 四叉树基本算法

    + makefile for linux

Plugins
    * GLRenderSystem
      + 渲染目标FBO实现版本

+++ Engine/AudioListener.cpp
@@ -12,13 +12,15 @@ AudioListener::AudioListener()
 : m_OldPosition(0.0f, 0.0f, 0.0f),
   m_VelocityFactor(0.0f)
 {}
+
 AudioListener::~AudioListener() {}
+
 void AudioListener::Update(unsigned long deltaTime)
 {
 	BaseSceneObject::Update(deltaTime);
 
 	// 更新收听者速度信息
-	Vector3f vel = (m_WorldTransform.GetPosition() - m_OldPosition) * 1000 / deltaTime * m_VelocityFactor;
+	Vector3f vel = (m_WorldTransform.GetPosition() - m_OldPosition) * 1000.0f / (float)deltaTime * m_VelocityFactor;
 	m_OldPosition = m_WorldTransform.GetPosition();
 	Engine::Instance().AudioSystem()->SetListenerTransform(this->WorldTransform(), vel);
 }
